refactor: split routes into their respective packages
